Split QueryCondition component out of QueryBuilder
Bug: T269133
